chore(utils): adjust codes
